changing for very longnumbers

+++ is_prime_prog/main.cpp
@@ -1,7 +1,7 @@
 #include<iostream>
 using namespace std;
 
-bool isPrime(int n)
+bool isPrime(unsigned long int n)
 {
     if (n <= 0 || n == 1)
         return false;
@@ -9,7 +9,7 @@ bool isPrime(int n)
         return true;
     if (n%2 == 0)
 	return false;
-    for (int i = 3; i< n/2; i+=2)
+    for (unsigned long int i = 3; i< n/2; i+=2)
         if (n % i == 0)
             return false;
     return true;
